requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
sqlparse==0.5.2
tomli==2.1.0
typing_extensions==4.12.2
